requirements.txt
streamlit>=1.32.0
sentence-transformers>=2.6.0
transformers>=4.38.0
torch>=2.0.0
fugashi>=1.3.0
ipadic>=1.0.0
unidic-lite>=1.0.8
mecab-python3>=1.0.6
requests>=2.31.0
beautifulsoup4>=4.12.0
scikit-learn>=1.4.0
lightgbm>=4.3.0
umap-learn>=0.5.6
matplotlib>=3.8.0
japanize-matplotlib>=1.1.3
seaborn>=0.13.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.12.0
plotly>=5.20.0
tqdm>=4.66.0
pdfplumber>=0.10.0
docx2txt>=0.8
PyPDF2>=3.0.1
